Generate a Load Profile

get_load_profile() returns load profiles in the form of quarter-hourly values (in Watt) that are standardized to an annual consumption of 1,000 kWh.

Usage

get_load_profile(
  profiles = c("H0", "G0", "G1", "G2", "G3", "G4", "G5", "G6", "L0", "L1", "L2"),
  start_date,
  end_date
)

Source

https://www.bdew.de/energie/standardlastprofile-strom/

https://www.bdew.de/media/documents/1999_Repraesentative-VDEW-Lastprofile.pdf

https://www.bdew.de/media/documents/2000131_Anwendung-repraesentativen_Lastprofile-Step-by-step.pdf

Arguments

profiles

name of load profiles, see 'Details'.

start_date

start date in ISO 8601 date format, required

end_date

end date in ISO 8601 date format, required

Value

A data.frame with four variables:

profile, character, load profile identifier

start_time, POSIXct / POSIXlt, start time

end_time, POSIXct / POSIXlt, end time

watts, numeric, electric power

Details

Supported profiles are:

H0: households (German: "Haushalte")

G0 to G6: commerce ("Gewerbe")

L0 to L2: agriculture ("Landwirtschaft")

Call get_load_profile_info() for more information about profiles.

The standard load profiles are differentiated according to winter, transitional period, summer as well as workday, Saturday, Sunday. See: https://www.bdew.de/media/documents/1999_Repraesentative-VDEW-Lastprofile.pdf for the methodology used to determine the profiles.

Periods are defined as:

summer: May 15 to September 14

winter: November 1 to March 20

transition: March 21 to May 14, and September 15 to October 31

Dates are mapped to one of three days:

saturday: Saturdays; Dec 24th and Dec 31th are considered a Saturday too, if they are not a Sunday

sunday: Sundays and all public holidays

workday: Monday to Friday

Note: The package supports nationwide, public holidays for Germany only. Those were retrieved from the nager.Date API and are listed below for 2024:

Jan 1: New Year's

Mar 29: Good Friday

Apr 1: Easter Monday

May 1: Labour Day

May 9: Ascension Day

May 20: Whit Monday

Oct 3: German Unity Day

Dec 25: Christmas Day:

Dec 26: Boxing Day

Examples

today <- Sys.Date()
get_load_profile("H0", today, today + 1) |> head()
#>   profile          start_time            end_time    watts
#> 1      H0 2023-11-26 00:00:00 2023-11-26 00:15:00 99.68435
#> 2      H0 2023-11-26 00:15:00 2023-11-26 00:30:00 92.39315
#> 3      H0 2023-11-26 00:30:00 2023-11-26 00:45:00 85.44373
#> 4      H0 2023-11-26 00:45:00 2023-11-26 01:00:00 78.72215
#> 5      H0 2023-11-26 01:00:00 2023-11-26 01:15:00 72.22843
#> 6      H0 2023-11-26 01:15:00 2023-11-26 01:30:00 66.30433
L <- get_load_profile(c("L0", "L1", "L2"), today, today + 1)
head(L)
#>   profile          start_time            end_time watts
#> 1      L0 2023-11-26 00:00:00 2023-11-26 00:15:00  68.3
#> 2      L0 2023-11-26 00:15:00 2023-11-26 00:30:00  66.0
#> 3      L0 2023-11-26 00:30:00 2023-11-26 00:45:00  64.3
#> 4      L0 2023-11-26 00:45:00 2023-11-26 01:00:00  63.0
#> 5      L0 2023-11-26 01:00:00 2023-11-26 01:15:00  62.1
#> 6      L0 2023-11-26 01:15:00 2023-11-26 01:30:00  61.4