Get the description load profiles according to the BDEW

Get the description load profiles according to the BDEW

Usage

get_load_profile_info(language = c("EN", "DE"))

Source

https://www.bdew.de/energie/standardlastprofile-strom/

Arguments

language

one of 'EN' (English), 'DE' (German)

Value

A data.frame

Examples

get_load_profile_info(language = "EN")
#>    profile                                                   description
#> 1       H0                                                     household
#> 2       G0                                           commerce in general
#> 3       G1                               commerce workday from 8am - 6pm
#> 4       G2        commerce with predominant consumption in evening hours
#> 5       G3                                            commerce continous
#> 6       G4                                             shop / barbershop
#> 7       G5                                         bakery with bakehouse
#> 8       G6                                              weekend business
#> 9       L0                                       agricultural in general
#> 10      L1 agricultural with dairy farming / part-time livestock farming
#> 11      L2                                 other agricultural businesses
#>                                                               comment
#> 1                                                                    
#> 2                                  weighted average of profiles G1-G6
#> 3  e.g. offices, doctor's office, workshop, administrative facilities
#> 4               e.g. sports club, fitness studio, evening restaurants
#> 5         e.g. cold storage warehouse, pumps, sewage treatment plants
#> 6                                                                    
#> 7                                                                    
#> 8                                                  e.g. movie theater
#> 9                                  weighted average of profiles L1-L2
#> 10                                                                   
#> 11                                                                   
get_load_profile_info(language = "DE")
#>    profile                                                         description
#> 1       H0                                                            Haushalt
#> 2       G0                                                   Gewerbe allgemein
#> 3       G1                                           Gewerbe werktags 8-18 Uhr
#> 4       G2  Gewerbe mit starkem bis überwiegendem Verbrauch in den Abendstunde
#> 5       G3                                                Gewerbe durchlaufend
#> 6       G4                                                      Laden/ Friseur
#> 7       G5                                              Bäckerei mit Backstube
#> 8       G6                                                    Wochenendbetrieb
#> 9       L0                                   Landwirtschaftsbetriebe allgemein
#> 10      L1 Landwirtschaftsbetriebe mit Milchwirtschaft/ Nebenerwerbs-Tierzucht
#> 11      L2                                      Übrige Landwirtschaftsbetriebe
#>                                                          comment
#> 1                                                               
#> 2                         Gewogener Mittelwert der Profile G1-G6
#> 3  z.B. Büros, Arztpraxen, Werkstätten, Verwaltungseinrichtungen
#> 4            z.B. Sportvereine, Fitnessstudios, Abendgaststätten
#> 5                           z.B. Kühlhäuser, Pumpen, Kläranlagen
#> 6                                                               
#> 7                                                               
#> 8                                                     z.B. Kinos
#> 9                     Gewogener Mittelwert der Profile L1 und L2
#> 10                                                              
#> 11